Move REPL related functions to new repl_functions.php
https://github.com/cakephp/cakephp/pull/14490#issuecomment-617809926
