+++ tests/TestCase/BasicsTest.php
@@ -22,13 +22,12 @@ use Cake\Collection\Collection;
 use Cake\Error\Debugger;
 use Cake\Event\EventManager;
 use Cake\Http\Response;
-use Cake\ORM\Table;
 use Cake\TestSuite\TestCase;
 
 require_once CAKE . 'basics.php';
 
 /**
- * BasicsTest class
+ * Basic functions tests
  */
 class BasicsTest extends TestCase
 {
@@ -426,17 +425,6 @@ EXPECTED;
         [$r, $expected] = [stackTrace($opts), \Cake\Error\Debugger::trace($opts)];
         $result = ob_get_clean();
         $this->assertSame($expected, $result);
-    }
-
-    /**
-     * Tests the table() function returns a table object
-     *
-     * @return void
-     */
-    public function testTable()
-    {
-        $table = \table('Table');
-        $this->assertInstanceOf(Table::class, $table);
     }
 
     /**
